Added first version of Blanc Evaluator Cleanup in Parent Evaluator and Document Changing labels to Double Creating abstract generic Mention Pair instance Modifications due to above Creating Logistic Regression Classifier

+++ g419-crete-api/src/main/java/g419/crete/api/classifier/factory/item/WekaJ48ClassifierItem.java
@@ -8,10 +8,10 @@ import java.util.List;
 import weka.classifiers.Classifier;
 import weka.core.Instance;
 
-public class WekaJ48ClassifierItem implements IClassifierFactoryItem<Classifier, Instance, Integer>{
+public class WekaJ48ClassifierItem implements IClassifierFactoryItem<Classifier, Instance, Double>{
 
 	@Override
-	public AbstractCreteClassifier<Classifier, Instance, Integer> createClassifier(List<String> features) {
+	public AbstractCreteClassifier<Classifier, Instance, Double> createClassifier(List<String> features) {
 		return new WekaDecisionTreesClassifier(features);
 	}
